Picture for tag and commit depends on AppSettings.
If this settings is true, commit and tag gpg info are retrieved in separated thread.
Revision.SignInfo is checked to choose the correct image.

Tag image, instead, depends on how many tag are there:
- zero tag -> column not visible
- one tag -> will be checked GPG result and image could be tag_ok or tag_error
- more than one tag -> tag_many image will be used

If many tag are present, the tagGpgMessage contains the result of all of them and will be displayed.